+++ tests/llm/prompt/test_snippet.py
@@ -1,5 +1,5 @@
 from unittest import TestCase
-from dandy.llm.prompt.snippet import TextSnippet, UnorderedListSnippet, OrderedListSnippet
+from dandy.llm.prompt import snippet
 
 
 class TestSnippet(TestCase):
@@ -23,20 +23,27 @@ class TestSnippet(TestCase):
             )
         ]
 
+    def test_object_source_snippet(self):
+        object_source = snippet.ObjectSourceSnippet(
+            object_module_name='dandy.llm.bot.llm_bot.BaseLlmBot'
+        )
+
+        self.assertEqual(object_source.to_str()[:10], '\nclass Bas')
+
     def test_text_snippet(self):
-        new_snippet = TextSnippet(text='Hello World')
+        new_snippet = snippet.TextSnippet(text='Hello World')
 
         self.assertEqual(new_snippet.to_str(), 'Hello World\n')
 
     def test_ordered_list_snippet(self):
-        snippet = OrderedListSnippet(
+        snippet.OrderedListSnippet(
             items=self.items
         )
 
         self.assertTrue(True)
 
     def test_unordered_list_snippet(self):
-        snippet = UnorderedListSnippet(
+        snippet.UnorderedListSnippet(
             items=self.items
         )